Add verbose logging options to story book
While not awesome as a default, it's sometimes useful to see items verbosely logged.

Seeing the events fire in realtime really helped clarify the difference between "controlled" and "uncontrolled" modes for me, for instance. Having this kind of thing set up for you is one of the great uses of storybook.