Scroll through autocomplete results on overflow
In the case of lots of autocomplete options, it's important that all
results are visible on the screen. This commit adds a default max height
for the autocomplete component, which can be overridden with a new
`autocompleteHeight` property. If the autocomplete results can't fit
within this height, the results will become scrollable.