Update README.md with classes prop example
As a relative newbie to Material UI getting my head around styling has taken some time. It wasn't immediately clear to me while reading the docs how to make a small change to some styles (despite the link to material-ui.com). Having a small example of using the classes prop (as proposed in my edit) should make it easier for the uninitiated!